requirements.txt
flask
flask-restful
flask-cors
imagekitio
pillow
requests
segno
uuid0